refactor: QueryProvider관련 설정 수정
